Changed definition of SERIAL_BUFFER_SIZE in cores/nRF5/RingBuffer.h to check if it is already defined.
Currently, SERIAL_BUFFER_SIZE is defined to have a specific value (64 in this case) regardless of whether it is already defined or not. This means that there is no way to change the size of the UART buffers without editting this file manually. This is already the case for the equivalent header for the SAMD core (https://github.com/adafruit/ArduinoCore-samd/blob/bdf5ac811733a9fdaf618fc61d8f5dcb8303ca00/cores/arduino/RingBuffer.h#L31-L33).